cloudops_agent/requirements.txt
boto3
langchain
openai
langgraph
python-dotenv
langchain-google-genai
google-generativeai
